pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "healthcare-nlp"
version = "0.1.0"
description = "Binary ADE (adverse drug event) sentence classifier fine-tuned from BERT, served via FastAPI on AWS ECS."
authors = [
    { name = "Felipe Díaz Jaramillo", email = "fdiazjaai@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch==2.10.0",
  "python-json-logger>=2.0,<3",
  "transformers==5.3.0",
  "accelerate==1.13.0",
  "datasets==4.7.0",
  "sentencepiece==0.2.1",
  "numpy==2.4.2",
  "scikit-learn==1.8.0",
  "pandas==3.0.1",
  "matplotlib==3.10.8",
  "tqdm==4.67.3",
  "wandb>=0.19,<1",
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
  "pydantic>=2.0,<3",
  "boto3>=1.34,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
  "sagemaker>=2,<3",
  "python-dotenv>=1.0,<2",
]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:importlib",
]
